move config file to config package
