+++ app/src/main/java/wzmame/pictour/activity/TestArea.java
@@ -8,8 +8,8 @@ import android.widget.Toast;
 import com.parse.Parse;
 import com.parse.ParseObject;
 
-import wzmame.pictour.ParseConfig;
 import wzmame.pictour.R;
+import wzmame.pictour.config.ParseConfig;
 
 public class TestArea extends AppCompatActivity {
